chore: extract code from scripts.js
